Moved WPF.Caches to - Services.Caches
