+++ Discreet-GUI/Services/Caches/SendTransactionCache.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace WPF.Caches
+namespace Services.Caches
 {
     public class SendTransactionCache
     {
