+++ Discreet-GUI/WPF/Hosted/DaemonActivatorService.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
+using Services.Caches;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -8,7 +9,6 @@ using System.Linq;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using WPF.Caches;
 using WPF.Services;
 
 namespace WPF.Hosted
