+++ Discreet-GUI/WPF/Startup.cs
@@ -3,6 +3,7 @@ using Avalonia.Controls.Notifications;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using Services.Caches;
 using Services.Daemon;
 using Services.Daemon.Services;
 using Services.ZMQ;
@@ -18,7 +19,6 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using WPF.Attributes;
-using WPF.Caches;
 using WPF.Factories.Navigation;
 using WPF.Factories.ViewModel;
 using WPF.Hosted;
