+++ Discreet-GUI/WPF/ValueConverters/AddressTypeToPublicPrivateStringConverter.cs
@@ -1,11 +1,11 @@
 ﻿using Avalonia.Data.Converters;
+using Services.Caches;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using WPF.Caches;
 
 namespace WPF.ValueConverters
 {
