+++ Discreet-GUI/WPF/Views/Account/AccountReceiveViewModel.cs
@@ -1,6 +1,8 @@
 ﻿using Avalonia.Controls;
 using Avalonia.Threading;
 using QRCoder;
+using Services.Caches;
+using Services.Extensions;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
@@ -9,8 +11,6 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using WPF.Caches;
-using WPF.ExtensionMethods;
 using WPF.ViewModels.Common;
 
 namespace WPF.Views.Account
