+++ Discreet-GUI/WPF/Views/Account/AccountSendViewModel.cs
@@ -1,16 +1,16 @@
 ﻿using Avalonia.Controls.Notifications;
 using ReactiveUI;
+using Services.Caches;
 using Services.Daemon;
 using Services.Daemon.Common;
 using Services.Daemon.Requests;
+using Services.Extensions;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Reactive;
 using System.Text;
 using System.Threading.Tasks;
-using WPF.Caches;
-using WPF.ExtensionMethods;
 using WPF.Factories.Navigation;
 using WPF.Services;
 using WPF.ViewModels.Common;
