+++ Discreet-GUI/WPF/Views/CreateWallet/WalletPasswordViewModel.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Reactive;
 using System.Text;
-using WPF.Caches;
+using Services.Caches;
 using WPF.Factories.Navigation;
 using WPF.Utility;
 using WPF.ViewModels.Common;
