+++ Discreet-GUI/WPF/Views/Layouts/Account/AccountLeftNavigationLayoutViewModel.cs
@@ -6,8 +6,8 @@ using System.Collections.ObjectModel;
 using System.Linq;
 using System.Reactive;
 using System.Text;
-using WPF.Caches;
-using WPF.ExtensionMethods;
+using Services.Caches;
+using Services.Extensions;
 using WPF.Factories.Navigation;
 using WPF.Stores.Navigation;
 using WPF.ViewModels.Common;
