+++ Discreet-GUI/WPF/Views/Layouts/DarkTitleBarLayoutSimpleViewModel.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using System.Reactive;
 using System.Text;
 using System.Threading.Tasks;
-using WPF.Caches;
+using Services.Caches;
 using WPF.Factories.Navigation;
 using WPF.Services;
 using WPF.Stores;
