+++ Discreet-GUI/WPF/Views/Modals/SelectWalletViewModel.cs
@@ -9,7 +9,7 @@ using System.Reactive.Concurrency;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using WPF.Caches;
+using Services.Caches;
 using WPF.Factories.Navigation;
 using WPF.Services;
 using WPF.ViewModels.Common;
