+++ Discreet-GUI/WPF/Views/Start/YourRecoveryPhraseViewModel.cs
@@ -10,7 +10,7 @@ using System.Linq;
 using System.Reactive;
 using System.Reactive.Concurrency;
 using System.Text;
-using WPF.Caches;
+using Services.Caches;
 using WPF.Factories.Navigation;
 using WPF.ViewModels.Common;
 using WPF.Views.CreateWallet;
